perf: improved separation of rendering concern
