+++ packages/morph/package.json
@@ -120,6 +120,7 @@
     "remark-smartypants": "^3.0.2",
     "rfdc": "^1.4.1",
     "shiki": "^1.29.2",
+    "sonner": "^2.0.3",
     "tailwind-merge": "^3.1.0",
     "tailwindcss-animate": "^1.0.7",
     "three": "^0.174.0",
